requirements.txt
# Core LLM functionality
google-generativeai>=0.3.0

# Environment variable management
python-dotenv>=1.0.0

# Vector operations and embeddings
numpy>=1.24.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Document processing
PyPDF2>=3.0.0